画面遷移時のバグ修正
